requirements.txt
absl-py==0.11.0
librosa==0.8.0
moviepy==1.0.3
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
tensorboard==2.4.0
torch==1.7.0
torchaudio==0.7.0
torchvision==0.8.1
tqdm==4.53.0
yacs==0.1.8
matplotlib
ffmpeg-python
scipy